pyproject.toml
[project]
name = "swirl_lm"
description = "Swirl-LM is a computational fluid dynamics simulation framework for low Mach number turbulent flows with variable density implemented with TensorFlow."
readme = "README.md"
# python 3.10 is used because it is the latest version that is supported by
# the public colab runtime.
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
  {name = "wqing", email="wqing@google.com"},
  {name = "anudhyan", email="anudhyan@google.com"},
  {name = "yifanchen", email="yifanchen@google.com"},
  ]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
]
keywords = []

# pip dependencies of the project
dependencies = [
    "absl-py",
    "attrs",
    "einops",
    "fancyflags",
    "h5py",
    "jax",
    "netCDF4",
    "numpy",
    "scipy",
    "six",
    "tensorflow-tpu",
]

# This is set automatically by flit using `swirl_lm.__version__`
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/google-research/swirl_lm"
repository = "https://github.com/google-research/swirl_lm"
# Other: `documentation`, `changelog`

[project.optional-dependencies]
# Development deps (unittest, linting, formating,...)
# Installed through `pip install .[dev]`
dev = [
    "pytest",
    "pytest-xdist",
    "pylint>=2.6.0",
    "yapf",
]

[build-system]
requires = ["flit_core >=3.5,<4"]
build-backend = "flit_core.buildapi"
